+++ models/UserContextModel.ts
@@ -1,4 +1,4 @@
-export interface Context {
+export default type UserContextModel = {
   timeFormat: number
   timezone: string
 }
